Add support for tracking transactions.
For safety, observable queries cannot be created nor subscribed to inside of a transaction. There's no safe way to limit the view of the database for these queries so we prevent it. A synchronous query method is provided for use inside a transaction instead.

This changes trigger semantics to emit multiple triggers at the same time and require queries to filter out the ones they don't care about. This allows us to have a transaction change multiple tables but only notify a query that is observing on all of those tables once.