Notify subscribers with a Query abstraction instead of a raw Cursor.
This approach has the benefit of not creating a resource which must be explicitly freed by subscribers. The disction means that methods like debounce and throttle which may ignore notifications are now fully usable at the expense of a single extra line in the subscriber. And since those operators are now available, the default throttling has been removed.