Update API to use SQLite support library.
This a pure port of existing usage. No new APIs were added to support things like SupportSQLiteQuery.

Because this support library does not use a normal open helper, interop with 1.x and 2.x is not possible. The recommended use will be to use 2.x and the interop library to get entirely to 2.x before switching to 3.x.